move ecosystem helpers to their own file
